requirements.txt
Django==4.2
gunicorn==20.1.0
whitenoise==6.4.0
psycopg2-binary==2.9.6
dj-database-url==1.3.0
djangorestframework==3.14.0
django-filter==23.2
python-dotenv==1.0.0